Move WZ QuickJS extensions
