+++ quickjs-wz-extensions/quickjs-limitedcontext.h
